+++ src/main/java/frc/robot/commands/ExampleSubsystem/ExampleCommand.java
@@ -2,20 +2,25 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ExampleSubsystem;
 
+import frc.robot.subsystems.ExampleSubsystem;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.RobotContainer;
-import frc.robot.subsystems.Susan;
 
-public class RotateSusan extends CommandBase {
-  /** Creates a new RotateSusan. */
-  Susan susan;
+/** An example command that uses an example subsystem. */
+public class ExampleCommand extends CommandBase {
+  @SuppressWarnings({ "PMD.UnusedPrivateField", "PMD.SingularField" })
+  private final ExampleSubsystem m_subsystem;
 
-  public RotateSusan(Susan p_susan) {
+  /**
+   * Creates a new ExampleCommand.
+   *
+   * @param subsystem The subsystem used by this command.
+   */
+  public ExampleCommand(ExampleSubsystem subsystem) {
+    m_subsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
-    susan = p_susan;
-    addRequirements(susan);
+    addRequirements(subsystem);
   }
 
   // Called when the command is initially scheduled.
@@ -26,16 +31,11 @@ public class RotateSusan extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    // check that susan is not over-extending itself
-    double rotation = RobotContainer.coDriverStick.getTwistAxis();
-    susan.setRotationSpeed(rotation);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    // stop rotating the susan
-    susan.setRotationSpeed(0.0);
   }
 
   // Returns true when the command should end.
