+++ src/main/java/frc/robot/commands/Turret/Susan/RotateSusan.java
@@ -2,20 +2,20 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Turret.Susan;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.RobotContainer;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.Susan;
 
-public class MoveDrivetrain extends CommandBase {
-  /** Creates a new MoveDrivetrain. */
-  Drivetrain drivetrain;
+public class RotateSusan extends CommandBase {
+  /** Creates a new RotateSusan. */
+  Susan susan;
 
-  public MoveDrivetrain(Drivetrain p_drivetrain) {
+  public RotateSusan(Susan p_susan) {
     // Use addRequirements() here to declare subsystem dependencies.
-    drivetrain = p_drivetrain;
-    addRequirements(drivetrain);
+    susan = p_susan;
+    addRequirements(susan);
   }
 
   // Called when the command is initially scheduled.
@@ -26,19 +26,16 @@ public class MoveDrivetrain extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    // read direction of joystick
-    double speed = RobotContainer.DriverStick.getArcadeMove();
-    double rotation = RobotContainer.DriverStick.getArcadeRotate();
-
-    // set motors to direction to joystick
-    drivetrain.setMovement(speed, rotation);
+    // check that susan is not over-extending itself
+    double rotation = RobotContainer.coDriverStick.getTwistAxis();
+    susan.setRotationSpeed(rotation);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    // stop spinning the wheels
-    drivetrain.setMovement(0, 0);
+    // stop rotating the susan
+    susan.setRotationSpeed(0.0);
   }
 
   // Returns true when the command should end.
